Consolidate/clean up wifi initialization
