+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CKIT_062S2_43012/cybsp.h
@@ -28,7 +28,7 @@
 #include "cybsp_types.h"
 #include "cybsp_core.h"
 #if defined(CYBSP_WIFI_CAPABLE)
-#include "cybsp_wifi_sdio.h"
+#include "cyhal_sdio.h"
 #endif
 #ifndef __MBED__
 #include "cybsp_retarget.h"
@@ -41,6 +41,30 @@ extern "C" {
 #endif
 
 
-#if defined(__cplusplus)
+/**
+* \addtogroup group_bsp_functions Functions
+* \{
+*/
+
+/**
+ * \brief Initialize all hardware on the board
+ * \returns CY_RSLT_SUCCESS if the board is sucessfully initialized, if there is
+ *          a problem initializing any hardware it returns an error code specific
+ *          to the hardware module that had a problem.
+ */
+cy_rslt_t cybsp_init(void);
+
+#if defined(CYBSP_WIFI_CAPABLE)
+/**
+ * \brief Get the initialized sdio object used for communicating with the WiFi Chip. 
+ * \note This function should only be called after cybsp_init();
+ * \returns The initialized sdio object.
+ */
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void);
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
+
+/** \} group_bsp_functions */
+
+#ifdef __cplusplus
 }
-#endif
+#endif /* __cplusplus */
