+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CKIT_062_BLE/cybsp.c
@@ -24,17 +24,26 @@
 
 #include <stdlib.h>
 #include "cybsp.h"
-#include "cycfg_system.h"
+#include "cyhal_utils.h"
+#include "cycfg.h"
 
 #if defined(__cplusplus)
 extern "C" {
 #endif
 
+#if defined(CYBSP_WIFI_CAPABLE)
+static cyhal_sdio_t sdio_obj;
+
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void)
+{
+    return &sdio_obj;
+}
+#endif
+
 cy_rslt_t cybsp_init(void)
 {
-    cy_rslt_t result;
-
-    result = cyhal_hwmgr_init();
+    /* Setup hardware manager to track resource usage then initialize all system (clock/power) board configuration */
+    cy_rslt_t result = cyhal_hwmgr_init();
     init_cycfg_system();
 
     if (CY_RSLT_SUCCESS == result)
@@ -42,7 +51,12 @@ cy_rslt_t cybsp_init(void)
         result = cybsp_register_sysclk_pm_callback();
     }
 
-#ifndef __MBED__
+#if defined(CYBSP_WIFI_CAPABLE)
+    /* Initialize SDIO interface. This must be done before other HAL API calls as some SDIO implementations require
+     * specific peripheral instances.
+     * NOTE: The full WiFi interface still needs to be initialized via cybsp_wifi_init_primary(). This is typically 
+     * done when starting up WiFi. 
+     */
     if (CY_RSLT_SUCCESS == result)
     {
         /* Initialize User LEDs */
@@ -68,8 +82,19 @@ cy_rslt_t cybsp_init(void)
             *   clock dividers */
             result = cybsp_retarget_init();
         }
+        /* Reserves: CYBSP_WIFI_SDIO, CYBSP_WIFI_SDIO_D0, CYBSP_WIFI_SDIO_D1, CYBSP_WIFI_SDIO_D2, CYBSP_WIFI_SDIO_D3
+         * CYBSP_WIFI_SDIO_CMD and CYBSP_WIFI_SDIO_CLK.
+         */
+        result = cyhal_sdio_init(
+                &sdio_obj,
+                CYBSP_WIFI_SDIO_CMD,
+                CYBSP_WIFI_SDIO_CLK,
+                CYBSP_WIFI_SDIO_D0,
+                CYBSP_WIFI_SDIO_D1,
+                CYBSP_WIFI_SDIO_D2,
+                CYBSP_WIFI_SDIO_D3);
     }
-#endif /* __MBED__ */
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
 
     /* CYHAL_HWMGR_RSLT_ERR_INUSE error code could be returned if any needed for BSP resource was reserved by
     *   user previously. Please review the Device Configurator (design.modus) and the BSP reservation list
