+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CKIT_062_BLE/cybsp.h
@@ -37,6 +37,21 @@
 extern "C" {
 #endif
 
-#if defined(__cplusplus)
+/**
+* \addtogroup group_bsp_functions Functions
+* \{
+*/
+#if defined(CYBSP_WIFI_CAPABLE)
+/**
+ * \brief Get the initialized sdio object used for communicating with the WiFi Chip. 
+ * \note This function should only be called after cybsp_init();
+ * \returns The initialized sdio object.
+ */
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void);
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
+
+/** \} group_bsp_functions */
+
+#ifdef __cplusplus
 }
-#endif
+#endif /* __cplusplus */
