+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CKIT_062_WIFI_BT/cybsp.h
@@ -28,7 +28,7 @@
 #include "cybsp_types.h"
 #include "cybsp_core.h"
 #if defined(CYBSP_WIFI_CAPABLE)
-#include "cybsp_wifi_sdio.h"
+#include "cyhal_sdio.h"
 #endif
 #ifndef __MBED__
 #include "cybsp_retarget.h"
@@ -47,6 +47,17 @@ extern cyhal_uart_t cybsp_debug_uart;
 extern cyhal_i2c_t cybsp_i2c;
 extern cyhal_rtc_t cybsp_rtc;
 
-#if defined(__cplusplus)
+#if defined(CYBSP_WIFI_CAPABLE)
+/**
+ * \brief Get the initialized sdio object used for communicating with the WiFi Chip. 
+ * \note This function should only be called after cybsp_init();
+ * \returns The initialized sdio object.
+ */
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void);
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
+
+/** \} group_bsp_functions */
+
+#ifdef __cplusplus
 }
-#endif
+#endif /* __cplusplus */
