+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CPROTO_064_SB/cybsp.c
@@ -23,17 +23,28 @@
 * limitations under the License.
 *******************************************************************************/
 
+#include <stdlib.h>
 #include "cybsp.h"
+#include "cyhal_utils.h"
 #include "cycfg.h"
 
 #if defined(__cplusplus)
 extern "C" {
 #endif
 
+#if defined(CYBSP_WIFI_CAPABLE)
+static cyhal_sdio_t sdio_obj;
+
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void)
+{
+    return &sdio_obj;
+}
+#endif
+
 cy_rslt_t cybsp_init(void)
 {
-    cy_rslt_t result = CY_RSLT_SUCCESS;
-
+    /* Setup hardware manager to track resource usage then initialize all system (clock/power) board configuration */
+    cy_rslt_t result = cyhal_hwmgr_init();
     init_cycfg_system();
     result = cybsp_register_sysclk_pm_callback();
 
@@ -52,8 +63,33 @@ cy_rslt_t cybsp_init(void)
             result = cybsp_retarget_init();
         }
     }
-#endif
-    
+
+#if defined(CYBSP_WIFI_CAPABLE)
+    /* Initialize SDIO interface. This must be done before other HAL API calls as some SDIO implementations require
+     * specific peripheral instances.
+     * NOTE: The full WiFi interface still needs to be initialized via cybsp_wifi_init_primary(). This is typically 
+     * done when starting up WiFi. 
+     */
+    if (CY_RSLT_SUCCESS == result)
+    {
+        /* Reserves: CYBSP_WIFI_SDIO, CYBSP_WIFI_SDIO_D0, CYBSP_WIFI_SDIO_D1, CYBSP_WIFI_SDIO_D2, CYBSP_WIFI_SDIO_D3
+         * CYBSP_WIFI_SDIO_CMD and CYBSP_WIFI_SDIO_CLK.
+         */
+        result = cyhal_sdio_init(
+                &sdio_obj,
+                CYBSP_WIFI_SDIO_CMD,
+                CYBSP_WIFI_SDIO_CLK,
+                CYBSP_WIFI_SDIO_D0,
+                CYBSP_WIFI_SDIO_D1,
+                CYBSP_WIFI_SDIO_D2,
+                CYBSP_WIFI_SDIO_D3);
+    }
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
+
+    /* CYHAL_HWMGR_RSLT_ERR_INUSE error code could be returned if any needed for BSP resource was reserved by
+     * user previously. Please review the Device Configurator (design.modus) and the BSP reservation list
+     * (cyreservedresources.list) to make sure no resources are reserved by both.
+     */
     return result;
 }
