+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CPROTO_064_SB/cybsp.h
@@ -30,7 +30,29 @@
 #ifndef __MBED__
 #include "cybsp_retarget.h"
 #endif /* __MBED__ */
+#if defined(CYBSP_WIFI_CAPABLE)
+#include "cyhal_sdio.h"
+#endif
 
 #if defined(__cplusplus)
-}
+extern "C" {
 #endif
+/**
+* \addtogroup group_bsp_functions Functions
+* \{
+*/
+
+#if defined(CYBSP_WIFI_CAPABLE)
+/**
+ * \brief Get the initialized sdio object used for communicating with the WiFi Chip. 
+ * \note This function should only be called after cybsp_init();
+ * \returns The initialized sdio object.
+ */
+cyhal_sdio_t* cybsp_get_wifi_sdio_obj(void);
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
+
+/** \} group_bsp_functions */
+
+#ifdef __cplusplus
+}
+#endif /* __cplusplus */
