+++ targets/TARGET_Cypress/TARGET_PSOC6/common/cy_network_buffer.h
@@ -32,7 +32,7 @@
 */
 
 #pragma once
-#if defined(TARGET_WHD)
+#if defined(CYBSP_WIFI_CAPABLE)
 
 #include <stdint.h>
 #include <stdbool.h>
@@ -176,6 +176,6 @@ void cy_network_process_ethernet_data(whd_interface_t interface, whd_buffer_t bu
 }
 #endif /* __cplusplus */
 
-#endif /* defined(TARGET_WHD) */
+#endif /* defined(CYBSP_WIFI_CAPABLE) */
 
 /** \} group_bsp_network_buffer */
