+++ targets/TARGET_Cypress/TARGET_PSOC6/common/cybsp_wifi.h
@@ -48,8 +48,11 @@ extern "C" {
 * \{
 */
 
-/** Initialization of the wifi driver failed. */
+/** Initialization of the WiFi driver failed. */
 #define CYBSP_RSLT_WIFI_INIT_FAILED (CY_RSLT_CREATE(CY_RSLT_TYPE_ERROR, CY_RSLT_MODULE_ABSTRACTION_BSP, 4))
+
+/** SDIO enumeration failed. */
+#define CYBSP_RSLT_WIFI_SDIO_ENUM_TIMEOUT (CY_RSLT_CREATE(CY_RSLT_TYPE_ERROR, CY_RSLT_MODULE_ABSTRACTION_BSP, 5))
 
 /** \} group_bsp_board_macros */
 
@@ -58,17 +61,21 @@ extern "C" {
 * \{
 */
 
-/** Initializes the primary interface for the wifi driver on the board. This function does the following
- * 1) Initializes the wifi driver.
- * 2) Turns on the WIFI chip.
+/** Initializes the primary interface for the WiFi driver on the board. This function
+ * does the following
+ * 1) Initializes the WiFi driver.
+ * 2) Turns on the WiFi chip.
+ *
+ * @note This function cannot be called multiple times. The user needs to first de-initialize
+ * before calling this function again.
  * 
  * @param[out] interface Interface to be initialized 
  * @return CY_RSLT_SUCCESS for successful initialization or error if initialization failed.
  */
 cy_rslt_t cybsp_wifi_init_primary(whd_interface_t* interface);
 
-/** This function initializes and adds a secondary interface to the wifi driver.
- *  @note This function does not initialize the wifi driver or turn on the wifi chip.
+/** This function initializes and adds a secondary interface to the WiFi driver.
+ *  @note This function does not initialize the WiFi driver or turn on the WiFi chip.
  * That is required to be done by the primary interface
  * 
  * @param[out] interface Interface to be initialized
@@ -76,6 +83,15 @@ cy_rslt_t cybsp_wifi_init_primary(whd_interface_t* interface);
  * @return CY_RSLT_SUCCESS for successful initialization or error if initialization failed.
  */
 cy_rslt_t cybsp_wifi_init_secondary(whd_interface_t* interface, whd_mac_t* mac_address);
+
+/** De-initializes all WiFi interfaces and the WiFi driver. This function does the following
+ * 1) Deinitializes all WiFi interfaces and WiFi driver.
+ * 2) Turns off the WiFi chip.
+ *
+ * @param[in] interface Interface to be de-initialized.
+ * @return CY_RSLT_SUCCESS for successful de-initialization or error if de-initialization failed.
+ */
+cy_rslt_t cybsp_wifi_deinit(whd_interface_t interface);
 
 /** Gets the wifi driver instance initialized by the driver. This should only be called
  * after the interface is initialized by cybsp_wifi_init_primary().
